pyproject.toml
[project]
name = "example-ollama-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.1",
    "langchain[google-genai]>=1.0.2",
    "langgraph-checkpoint-postgres>=3.0.0",
    "ollama>=0.6.0",
    "psycopg-binary>=3.2.12",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
